+++ src/app/app.module.ts
@@ -4,12 +4,13 @@ import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 
-import { ProductService } from './services/products.service';
-import { CartService } from './services/cart.service';
+import { CartService } from './modules/cart/services/cart.service';
 import { CartModule } from './modules/cart/cart.module';
 import { OrdersModule } from './modules/orders/orders.module';
 import { ProductModule } from './modules/product/product.module';
 import { SharedModule } from './modules/shared/shared.module';
+import { ProductService } from './modules/product/services/products.service';
+import { FirstModule } from './modules/first/first.module';
 
 @NgModule({
   declarations: [
@@ -21,6 +22,7 @@ import { SharedModule } from './modules/shared/shared.module';
     CartModule,
     OrdersModule,
     ProductModule,
+    FirstModule,
     SharedModule
   ],
   providers: [ProductService, CartService],