+++ src/app/models/config.model.ts
@@ -1,0 +1,6 @@
+
+export class ConfigModel {
+    id: string;
+    login: string;
+    email: string;
+  }