+++ src/app/modules/cart/services/cart.service.ts
@@ -1,12 +1,13 @@
 import { Injectable } from '@angular/core';
-import { BoughtProductModel } from '../models/boughtProduct.model';
-import { ProductModel } from '../models/product.model';
+import { BoughtProductModel } from '../../../models/bought-product.model';
+import { ProductModel } from '../../../models/product.model';
 
 @Injectable()
 export class CartService {
     private products: Array<BoughtProductModel> = new Array<BoughtProductModel>();
     private counter = 0;
     private totalAmount = 0;
+    private isEmptyList = true;
 
     addProduct(product: ProductModel): void {
         if (this.contains(product)) {
@@ -15,23 +16,19 @@ export class CartService {
         } else {
             this.products.push(new BoughtProductModel(product, 1));
         }
-
-        this.counter++;
-        this.totalAmount += product.price;
+        this.updateCartData();
     }
 
     removeProduct(product: ProductModel): void {
         if (this.contains(product)) {
             const boughtProduct = this.getBoughtProductByProduct(product);
-            if (boughtProduct.quantity == 1) {
+            if (boughtProduct.quantity === 1) {
                 const index = this.products.indexOf(boughtProduct);
                 this.products.splice(index, 1);
             } else {
                 boughtProduct.quantity--;
             }
-
-            this.counter--;
-            this.totalAmount -= product.price;
+            this.updateCartData();
         }
     }
 
@@ -42,19 +39,16 @@ export class CartService {
 
         if (this.contains(product)) {
             const boughtProduct = this.getBoughtProductByProduct(product);
-            const difference = quantity - boughtProduct.quantity;
 
-            if (quantity == 0) {
+            if (quantity === 0) {
                 const index = this.products.indexOf(boughtProduct);
                 this.products.splice(index, 1);
             }
             else {
                 boughtProduct.quantity = quantity;
             }
-
-            this.counter += difference;
-            this.totalAmount += product.price * difference;
         }
+        this.updateCartData();
     }
 
     getProducts(): Array<BoughtProductModel> {
@@ -65,19 +59,43 @@ export class CartService {
         return this.getBoughtProductByProduct(product) != null;
     }
 
-    get count(): number {
+    removeAllProducts(): void {
+        this.products = new Array<BoughtProductModel>();
+        this.updateCartData();
+    }
+
+    get totalQuantity(): number {
         return this.counter;
     }
 
-    get amount(): number {
+    get totalSum(): number {
         return this.totalAmount;
+    }
+
+    get isEmpty(): boolean {
+        return this.isEmptyList;
+    }
+
+    private updateCartData(): void {
+        this.counter = 0;
+        this.totalAmount = 0;
+        this.products.forEach(element => {
+            this.counter += element.quantity;
+            this.totalAmount += element.quantity * element.product.price;
+        });
+
+        if (this.counter == 0){
+            this.isEmptyList = true;
+        }else{
+            this.isEmptyList = false;
+        }
     }
 
     private getBoughtProductByProduct(product: ProductModel): BoughtProductModel {
         let boughtProduct = null;
 
         this.products.forEach(x => {
-            if (x.product == product) {
+            if (x.product === product) {
                 boughtProduct = x;
             }
         });