+++ src/app/modules/shared/shared.module.ts
@@ -1,14 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './highlight.directive';
+import { ClickDirective } from './click.directive';
 
 
 
 @NgModule({
-  declarations: [HighlightDirective],
+  declarations: [HighlightDirective, ClickDirective],
   imports: [
     CommonModule,
   ],
-  exports: [HighlightDirective]
+  exports: [HighlightDirective, ClickDirective]
 })
 export class SharedModule { }