+++ changelog.md
@@ -1,1 +1,10 @@
-Всі зміни опишу 29 жовтня коли піду у відпустку Пробачте зараз просто нема часу
+
+Services
+1. implemented interface of CartService
+2. created LocalStorageService
+3. created ConfigOptionsService
+4. created ConstantsService
+5. created GeneratorService
+6. created FirstComponent
+7. injected services into FirstComponent
+8. created click directive